+++ __tests__/index.ts
@@ -5,23 +5,6 @@ import memoize from '../src';
 import { isSameValueZero } from '../src/utils';
 
 describe('memoize', () => {
-  it('will return the function if already memoized', () => {
-    const fn = () => {};
-
-    fn.isMemoized = true;
-
-    const memoized = memoize(fn);
-
-    expect(memoized).toBe(fn);
-  });
-
-  it('will throw an error if not a function', () => {
-    const fn = 123;
-
-    // @ts-ignore
-    expect(() => memoize(fn)).toThrow();
-  })
-
   it('will return the memoized function', () => {
     let callCount = 0;
 
@@ -36,12 +19,12 @@ describe('memoize', () => {
 
     const memoized = memoize(fn);
 
-    expect(memoized.cache).toEqual({
+    expect(memoized.cache.snapshot).toEqual({
       keys: [],
       size: 0,
       values: [],
     });
-    expect(memoized.cacheSnapshot).toEqual({
+    expect(memoized.cache.snapshot).toEqual({
       keys: [],
       size: 0,
       values: [],
@@ -68,7 +51,7 @@ describe('memoize', () => {
 
     expect(callCount).toEqual(1);
 
-    expect(memoized.cache).toEqual({
+    expect(memoized.cache.snapshot).toEqual({
       keys: [['one', 'two']],
       size: 1,
       values: [
@@ -95,13 +78,13 @@ describe('memoize', () => {
 
     const memoized = memoize(fn, { maxSize });
 
-    expect(memoized.cache).toEqual({
+    expect(memoized.cache.snapshot).toEqual({
       keys: [],
       size: 0,
       values: [],
     });
 
-    expect(memoized.cacheSnapshot).toEqual({
+    expect(memoized.cache.snapshot).toEqual({
       keys: [],
       size: 0,
       values: [],
@@ -138,7 +121,7 @@ describe('memoize', () => {
 
     expect(callCount).toEqual(4);
 
-    expect(memoized.cache).toEqual({
+    expect(memoized.cache.snapshot).toEqual({
       keys: [['three', 'four'], ['two', 'three'], ['four', 'five']],
       size: 3,
       values: [
@@ -196,7 +179,7 @@ describe('memoize', () => {
 
     expect(callCount).toEqual(1);
 
-    expect(memoized.cache).toEqual({
+    expect(memoized.cache.snapshot).toEqual({
       keys: [
         [{ deep: { value: 'value' } }, { other: { deep: { value: 'value' } } }],
       ],
@@ -250,7 +233,7 @@ describe('memoize', () => {
 
     expect(callCount).toEqual(1);
 
-    expect(memoized.cache).toEqual({
+    expect(memoized.cache.snapshot).toEqual({
       keys: [['[{"one":"one"},null]']],
       size: 1,
       values: [
@@ -311,7 +294,7 @@ describe('memoize', () => {
 
     expect(callCount).toEqual(1);
 
-    expect(memoized.cache).toEqual({
+    expect(memoized.cache.snapshot).toEqual({
       keys: [
         [
           {
@@ -351,14 +334,14 @@ describe('memoize', () => {
 
     memoized('foo').catch(spy);
 
-    expect(memoized.cacheSnapshot.keys.length).toEqual(1);
-    expect(memoized.cacheSnapshot.values.length).toEqual(1);
+    expect(memoized.cache.snapshot.keys.length).toEqual(1);
+    expect(memoized.cache.snapshot.values.length).toEqual(1);
 
     await new Promise((resolve: Function) => {
       setTimeout(resolve, timeout + 50);
     });
 
-    expect(memoized.cacheSnapshot).toEqual({
+    expect(memoized.cache.snapshot).toEqual({
       keys: [],
       size: 0,
       values: [],
@@ -409,24 +392,8 @@ describe('memoize', () => {
 
     expect(onCacheChange).toHaveBeenCalledTimes(1);
     expect(onCacheChange).toHaveBeenCalledWith(
-      {
-        keys: [['foo', 'bar']],
-        size: 1,
-        values: [
-          {
-            one: 'foo',
-            two: 'bar',
-          },
-        ],
-      },
-      {
-        maxSize,
-        onCacheChange,
-        isEqual: isSameValueZero,
-        isMatchingKey: undefined,
-        isPromise: false,
-        transformKey: undefined,
-      },
+      memoized.cache,
+      memoized.options,
       memoized,
     );
 
@@ -436,28 +403,8 @@ describe('memoize', () => {
 
     expect(onCacheChange).toHaveBeenCalledTimes(1);
     expect(onCacheChange).toHaveBeenCalledWith(
-      {
-        keys: [['bar', 'foo'], ['foo', 'bar']],
-        size: 2,
-        values: [
-          {
-            one: 'bar',
-            two: 'foo',
-          },
-          {
-            one: 'foo',
-            two: 'bar',
-          },
-        ],
-      },
-      {
-        maxSize,
-        onCacheChange,
-        isEqual: isSameValueZero,
-        isMatchingKey: undefined,
-        isPromise: false,
-        transformKey: undefined,
-      },
+      memoized.cache,
+      memoized.options,
       memoized,
     );
 
@@ -473,28 +420,8 @@ describe('memoize', () => {
 
     expect(onCacheChange).toHaveBeenCalledTimes(1);
     expect(onCacheChange).toHaveBeenCalledWith(
-      {
-        keys: [['foo', 'bar'], ['bar', 'foo']],
-        size: 2,
-        values: [
-          {
-            one: 'foo',
-            two: 'bar',
-          },
-          {
-            one: 'bar',
-            two: 'foo',
-          },
-        ],
-      },
-      {
-        maxSize,
-        onCacheChange,
-        isEqual: isSameValueZero,
-        isMatchingKey: undefined,
-        isPromise: false,
-        transformKey: undefined,
-      },
+      memoized.cache,
+      memoized.options,
       memoized,
     );
 
@@ -545,28 +472,8 @@ describe('memoize', () => {
 
     expect(onCacheHit).toHaveBeenCalledTimes(1);
     expect(onCacheHit).toHaveBeenCalledWith(
-      {
-        keys: [['bar', 'foo'], ['foo', 'bar']],
-        size: 2,
-        values: [
-          {
-            one: 'bar',
-            two: 'foo',
-          },
-          {
-            one: 'foo',
-            two: 'bar',
-          },
-        ],
-      },
-      {
-        maxSize,
-        onCacheHit,
-        isEqual: isSameValueZero,
-        isMatchingKey: undefined,
-        isPromise: false,
-        transformKey: undefined,
-      },
+      memoized.cache,
+      memoized.options,
       memoized,
     );
 
@@ -576,28 +483,8 @@ describe('memoize', () => {
 
     expect(onCacheHit).toHaveBeenCalledTimes(1);
     expect(onCacheHit).toHaveBeenCalledWith(
-      {
-        keys: [['foo', 'bar'], ['bar', 'foo']],
-        size: 2,
-        values: [
-          {
-            one: 'foo',
-            two: 'bar',
-          },
-          {
-            one: 'bar',
-            two: 'foo',
-          },
-        ],
-      },
-      {
-        maxSize,
-        onCacheHit,
-        isEqual: isSameValueZero,
-        isMatchingKey: undefined,
-        isPromise: false,
-        transformKey: undefined,
-      },
+      memoized.cache,
+      memoized.options,
       memoized,
     );
 
@@ -607,28 +494,8 @@ describe('memoize', () => {
 
     expect(onCacheHit).toHaveBeenCalledTimes(1);
     expect(onCacheHit).toHaveBeenCalledWith(
-      {
-        keys: [['foo', 'bar'], ['bar', 'foo']],
-        size: 2,
-        values: [
-          {
-            one: 'foo',
-            two: 'bar',
-          },
-          {
-            one: 'bar',
-            two: 'foo',
-          },
-        ],
-      },
-      {
-        maxSize,
-        onCacheHit,
-        isEqual: isSameValueZero,
-        isMatchingKey: undefined,
-        isPromise: false,
-        transformKey: undefined,
-      },
+      memoized.cache,
+      memoized.options,
       memoized,
     );
   });
@@ -672,24 +539,8 @@ describe('memoize', () => {
 
     expect(onCacheAdd).toHaveBeenCalledTimes(1);
     expect(onCacheAdd).toHaveBeenCalledWith(
-      {
-        keys: [['foo', 'bar']],
-        size: 1,
-        values: [
-          {
-            one: 'foo',
-            two: 'bar',
-          },
-        ],
-      },
-      {
-        maxSize,
-        onCacheAdd,
-        isEqual: isSameValueZero,
-        isMatchingKey: undefined,
-        isPromise: false,
-        transformKey: undefined,
-      },
+      memoized.cache,
+      memoized.options,
       memoized,
     );
 
@@ -699,28 +550,8 @@ describe('memoize', () => {
 
     expect(onCacheAdd).toHaveBeenCalledTimes(1);
     expect(onCacheAdd).toHaveBeenCalledWith(
-      {
-        keys: [['bar', 'foo'], ['foo', 'bar']],
-        size: 2,
-        values: [
-          {
-            one: 'bar',
-            two: 'foo',
-          },
-          {
-            one: 'foo',
-            two: 'bar',
-          },
-        ],
-      },
-      {
-        maxSize,
-        onCacheAdd,
-        isEqual: isSameValueZero,
-        isMatchingKey: undefined,
-        isPromise: false,
-        transformKey: undefined,
-      },
+      memoized.cache,
+      memoized.options,
       memoized,
     );
 
@@ -743,11 +574,14 @@ describe('memoize', () => {
 
   type Dictionary<Type> = {
     [key: string]: Type;
-  }
+  };
 
   test('if recursive calls to self will be respected at runtime', () => {
     const calc = memoize(
-      (object: { [key: string]: any }, metadata: { c: number }): Dictionary<any> =>
+      (
+        object: { [key: string]: any },
+        metadata: { c: number },
+      ): Dictionary<any> =>
         Object.keys(object).reduce((totals: { [key: string]: number }, key) => {
           if (Array.isArray(object[key])) {
             totals[key] = object[key].map(
@@ -797,5 +631,27 @@ describe('memoize', () => {
     const result2 = calc(data, metadata);
 
     expect(result1).toEqual(result2);
+  });
+
+  it('will re-memoize the function with merged options if already memoized', () => {
+    const fn = () => {};
+
+    const maxSize = 5;
+    const isEqual = () => true;
+
+    const memoized = memoize(fn, { maxSize });
+
+    const reMemoized = memoize(memoized, { isEqual });
+
+    expect(reMemoized).not.toBe(memoized);
+    expect(reMemoized.options.maxSize).toBe(maxSize);
+    expect(reMemoized.options.isEqual).toBe(isEqual);
+  });
+
+  it('will throw an error if not a function', () => {
+    const fn = 123;
+
+    // @ts-ignore
+    expect(() => memoize(fn)).toThrow();
   });
 });