+++ index.d.ts
@@ -5,18 +5,26 @@ type Dictionary<Type> = {
 
 declare namespace MicroMemoize {
   export type Key = any[];
+  export type Value = any;
 
   export type RawKey = Key | IArguments;
 
-  export type Keys = Key[];
-
-  export type Values = any[];
-
-  export type Cache = {
-    keys: Keys;
-    size: number;
-    values: Values;
-  };
+  export interface Cache {
+    canTransformKey: boolean;
+    getKeyIndex: KeyIndexGetter;
+    keys: Key[];
+    options: Options;
+    shouldCloneArguments: boolean;
+    shouldUpdateOnAdd: boolean;
+    shouldUpdateOnChange: boolean;
+    shouldUpdateOnHit: boolean;
+    snapshot: {
+      keys: Key[];
+      size: number;
+      values: Value[];
+    };
+    values: Value[];
+  }
 
   export type EqualityComparator = (object1: any, object2: any) => boolean;
 
@@ -30,12 +38,7 @@ declare namespace MicroMemoize {
 
   export type KeyTransformer = (args: RawKey) => Key;
 
-  export type KeyIndexGetter = (allKeys: Keys, keyToMatch: RawKey) => number;
-
-  export type AsyncCacheUpdater = (
-    cache: Cache,
-    memoized: Memoized<Function>,
-  ) => void;
+  export type KeyIndexGetter = (allKeys: Key[], keyToMatch: RawKey) => number;
 
   export type StandardOptions = {
     isEqual?: EqualityComparator;
@@ -50,11 +53,11 @@ declare namespace MicroMemoize {
 
   export type Options = StandardOptions & Dictionary<any>;
 
-  export type Memoized<Fn extends Function> = Fn & {
-    [key: string]: any;
-    cache?: Cache;
-    cacheSnapshot?: Cache;
-    isMemoized?: boolean;
-    options?: Options;
-  };
+  export type Memoized<Fn extends Function> = Fn &
+    Dictionary<any> & {
+      cache?: Cache;
+      fn: Fn;
+      isMemoized?: boolean;
+      options?: Options;
+    };
 }