+++ src/index.ts
@@ -1,32 +1,27 @@
+// cache
+import { Cache } from './Cache';
+
 // types
-import { Cache, Memoized, Keys, StandardOptions, Values } from './types';
+import { Memoized, StandardOptions } from './types';
 
 // utils
 import {
-  createGetKeyIndex,
-  createUpdateAsyncCache,
   getCustomOptions,
-  isFunction,
   isMemoized,
   isSameValueZero,
   mergeOptions,
-  orderByLru,
+  slice,
 } from './utils';
 
-const slice = Function.prototype.bind.call(
-  Function.prototype.call,
-  Array.prototype.slice,
-);
-
 function createMemoizedFunction<Fn extends Function>(
-  fn: Fn,
+  fn: Fn | Memoized<Fn>,
   options: StandardOptions = {},
 ): Memoized<Fn> {
   if (isMemoized(fn)) {
-    return fn;
+    return createMemoizedFunction(fn.fn, mergeOptions(fn.options, options));
   }
 
-  if (!isFunction(fn)) {
+  if (typeof fn !== 'function') {
     throw new TypeError('You must pass a function to `memoize`.');
   }
 
@@ -41,95 +36,80 @@ function createMemoizedFunction<Fn extends Function>(
     transformKey,
   }: StandardOptions = options;
 
-  const normalizedOptions = mergeOptions(getCustomOptions(options), {
-    isEqual,
-    isMatchingKey,
-    isPromise,
-    maxSize,
-    onCacheAdd,
-    onCacheChange,
-    onCacheHit,
-    transformKey,
-  });
-
-  const getKeyIndex = createGetKeyIndex(normalizedOptions);
-  const updateAsyncCache = createUpdateAsyncCache(normalizedOptions);
-
-  const keys: Keys = [];
-  const values: Values = [];
-
-  const cache: Cache = {
-    keys,
-    get size() {
-      return cache.keys.length;
+  const normalizedOptions = mergeOptions(
+    {
+      isEqual,
+      isMatchingKey,
+      isPromise,
+      maxSize,
+      onCacheAdd,
+      onCacheChange,
+      onCacheHit,
+      transformKey,
     },
+    getCustomOptions(options),
+  );
+
+  const cache = new Cache(normalizedOptions);
+
+  const {
+    keys,
     values,
-  };
-
-  const canTransformKey = typeof transformKey === 'function';
-
-  const shouldCloneArguments = !!(transformKey || isMatchingKey);
-
-  const shouldUpdateOnAdd = typeof onCacheAdd === 'function';
-  const shouldUpdateOnChange = typeof onCacheChange === 'function';
-  const shouldUpdateOnHit = typeof onCacheHit === 'function';
+    canTransformKey,
+    shouldCloneArguments,
+    shouldUpdateOnAdd,
+    shouldUpdateOnChange,
+    shouldUpdateOnHit,
+  } = cache;
 
   // @ts-ignore
   const memoized: Memoized<Fn> = function memoized() {
-    const normalizedArgs = shouldCloneArguments
-      ? slice(arguments, 0)
-      : arguments;
-    const key = canTransformKey ? transformKey(normalizedArgs) : normalizedArgs;
-    const keyIndex = keys.length ? getKeyIndex(keys, key) : -1;
+    let key = shouldCloneArguments ? slice(arguments, 0) : arguments;
+
+    if (canTransformKey) {
+      key = transformKey(key);
+    }
+
+    const keyIndex = keys.length ? cache.getKeyIndex(key) : -1;
 
     if (keyIndex !== -1) {
-      shouldUpdateOnHit && onCacheHit(cache, normalizedOptions, memoized);
+      if (shouldUpdateOnHit) {
+        onCacheHit(cache, normalizedOptions, memoized);
+      }
 
       if (keyIndex) {
-        orderByLru(cache, keys[keyIndex], values[keyIndex], keyIndex, maxSize);
+        cache.orderByLru(keys[keyIndex], values[keyIndex], keyIndex);
 
-        shouldUpdateOnChange &&
+        if (shouldUpdateOnChange) {
           onCacheChange(cache, normalizedOptions, memoized);
+        }
       }
     } else {
       const newValue = fn.apply(this, arguments);
       const newKey = shouldCloneArguments ? key : slice(arguments, 0);
 
-      orderByLru(cache, newKey, newValue, keys.length, maxSize);
+      cache.orderByLru(newKey, newValue, keys.length);
 
-      isPromise && updateAsyncCache(cache, memoized);
+      if (isPromise) {
+        cache.updateAsyncCache(memoized);
+      }
 
-      shouldUpdateOnAdd && onCacheAdd(cache, normalizedOptions, memoized);
-      shouldUpdateOnChange && onCacheChange(cache, normalizedOptions, memoized);
+      if (shouldUpdateOnAdd) {
+        onCacheAdd(cache, normalizedOptions, memoized);
+      }
+
+      if (shouldUpdateOnChange) {
+        onCacheChange(cache, normalizedOptions, memoized);
+      }
     }
 
     return values[0];
   };
 
-  Object.defineProperties(memoized, {
-    cache: {
-      configurable: true,
-      value: cache,
-    },
-    cacheSnapshot: {
-      configurable: true,
-      get() {
-        return {
-          keys: slice(cache.keys, 0),
-          size: cache.size,
-          values: slice(cache.values, 0),
-        };
-      },
-    },
-    isMemoized: {
-      configurable: true,
-      value: true,
-    },
-    options: {
-      configurable: true,
-      value: normalizedOptions,
-    },
-  });
+  memoized.cache = cache;
+  memoized.fn = fn;
+  memoized.isMemoized = true as const;
+  memoized.options = normalizedOptions;
 
   return memoized;
 }