+++ src/types.ts
@@ -4,18 +4,26 @@ export type Dictionary<Type> = {
 };
 
 export type Key = any[];
+export type Value = any;
 
 export type RawKey = Key | IArguments;
 
-export type Keys = Key[];
-
-export type Values = any[];
-
-export type Cache = {
-  keys: Keys;
-  size: number;
-  values: Values;
-};
+export interface Cache {
+  canTransformKey: boolean;
+  getKeyIndex: KeyIndexGetter;
+  keys: Key[];
+  options: Options;
+  shouldCloneArguments: boolean;
+  shouldUpdateOnAdd: boolean;
+  shouldUpdateOnChange: boolean;
+  shouldUpdateOnHit: boolean;
+  snapshot: {
+    keys: Key[];
+    size: number;
+    values: Value[];
+  };
+  values: Value[];
+}
 
 export type EqualityComparator = (object1: any, object2: any) => boolean;
 
@@ -29,12 +37,7 @@ export type CacheModifiedHandler = (
 
 export type KeyTransformer = (args: RawKey) => Key;
 
-export type KeyIndexGetter = (allKeys: Keys, keyToMatch: RawKey) => number;
-
-export type AsyncCacheUpdater = (
-  cache: Cache,
-  memoized: Memoized<Function>,
-) => void;
+export type KeyIndexGetter = (keyToMatch: RawKey) => number;
 
 export type StandardOptions = {
   isEqual?: EqualityComparator;
@@ -49,10 +52,10 @@ export type StandardOptions = {
 
 export type Options = StandardOptions & Dictionary<any>;
 
-export type Memoized<Fn extends Function> = Fn & {
-  [key: string]: any;
-  cache?: Cache;
-  cacheSnapshot?: Cache;
-  isMemoized?: boolean;
-  options?: Options;
-};
+export type Memoized<Fn extends Function> = Fn &
+  Dictionary<any> & {
+    cache: Cache;
+    fn: Fn;
+    isMemoized: true;
+    options: Options;
+  };